Authentication fix

+++ src/main/java/com/online/store/repository/UserRepository.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByUsername(String username);
-    Optional<User> findById(Long id);
+    Optional<User> findById(int id);
     boolean existsByUsername(String username);
     boolean existsByEmail(String email);
 }
